mistral/requirements.txt
ctransformers
simple-parsing
sentence_transformers
psycopg2-binary
pgvector
langchain